backend_new/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.9.4
aiosignal==1.3.2
aiosqlite==0.19.0
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.21.3
anyio==4.8.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
atlassian-python-api==3.41.4
attrs==25.1.0
azure-core==1.38.3
azure-devops==7.1.0b3
azure-identity==1.15.0
babel==2.17.0
bcrypt==4.2.1
beautifulsoup4==4.14.3
bitarray==3.8.0
bleach==6.2.0
blinker==1.9.0
boltons==21.0.0
boto3==1.33.6
botocore==1.33.13
bracex==2.6
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.1
click-option-group==0.5.9
colorama==0.4.6
contourpy==1.2.1
coverage==7.13.2
cryptography==44.0.1
cycler==0.12.1
cytoolz==1.1.0
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.2.0
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
dynaconf==3.2.4
ecdsa==0.19.1
email-validator==2.3.0
environs==14.1.1
eth-account==0.10.0
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
exceptiongroup==1.2.2
executing==2.2.0
face==26.0.0
fastapi==0.111.0
fastapi-cli==0.0.24
fastapi_cors==0.0.6
fastjsonschema==2.21.1
filelock==3.25.2
filetype==1.2.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2026.2.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.41
glom==25.12.0
google-ai-generativelanguage==0.6.18
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-aiplatform==1.38.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==2.10.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gotrue==2.11.3
greenlet==3.1.1
grpc-google-iam-v1==0.14.3
grpcio==1.78.1
grpcio-status==1.62.3
gunicorn==22.0.0
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hexbytes==0.3.1
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.7
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==1.7.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.1
iniconfig==2.3.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.2
jiter==0.9.0
jmespath==1.0.1
joblib==1.3.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
langchain==0.3.22
langchain-anthropic==0.3.0
langchain-community==0.3.20
langchain-core==0.3.49
langchain-google-genai==2.1.2
langchain-openai==0.3.12
langchain-text-splitters==0.3.7
langsmith==0.3.45
libclang==18.1.1
litellm==1.43.13
loguru==0.7.2
lru-dict==1.2.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.2
mcp==1.23.3
mdurl==0.1.2
mistune==3.1.2
ml-dtypes==0.4.1
msal==1.35.1
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.3.0
namex==0.0.8
narwhals==1.27.1
nest-asyncio==1.6.0
numpy==1.26.3
oauthlib==3.3.1
openai==1.40.6
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
opt_einsum==3.4.0
orjson==3.10.16
overrides==7.7.0
packaging==26.0
pandas==2.1.4
pandocfilters==1.5.1
paramiko==3.5.1
parsimonious==0.10.0
parso==0.8.4
passlib==1.7.4
peewee==3.19.0
pillow==10.4.0
platformdirs==4.3.6
plotly==6.0.0
pluggy==1.6.0
postgrest==0.19.3
pr-agent==0.2.4
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.27.1
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
py==1.11.0
py-ecc==8.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.8.2
pydantic-settings==2.12.0
pydantic_core==2.20.1
PyGithub==1.59.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.1.2
pytest==7.4.0
pytest-asyncio==0.23.3
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-gitlab==3.15.0
python-jose==3.3.0
python-json-logger==3.2.1
python-multipart==0.0.22
pytz==2025.1
pyunormalize==17.0.0
pywin32==311; platform_system == "Windows"
pywinpty==2.0.15; platform_system == "Windows"
PyYAML==6.0.1
pyzmq==26.2.1
realtime==2.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.3.3
rich-toolkit==0.19.7
rlp==4.1.0
rpds-py==0.30.0
rsa==4.9
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.14
s3transfer==0.8.2
scapy==2.6.1
scikit-learn==1.4.0
scipy==1.15.2
seaborn==0.13.1
semantic-version==2.10.0
semgrep==1.152.0
Send2Trash==1.8.3
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.25
sse-starlette==3.2.0
stack-data==0.6.3
starlette==0.37.2
starlette-context==0.3.6
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
telnetlib3==2.0.4
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.0.2
toolz==1.1.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.24.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
ujson==5.8.0
uri-template==1.3.0
urllib3==2.0.7
uuid_utils==0.14.1
uvicorn==0.22.0
watchfiles==1.1.1
wcmatch==8.5.2
wcwidth==0.2.13
web3==6.15.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.13
win32_setctime==1.2.0
wrapt==1.17.2
xgboost==2.0.3
yarl==1.18.3
zipp==3.23.0
zstandard==0.23.0
